+++ src/cli/upgrade_banner.js
